Moved encode and decode of user token outside of the user class
