+++ src/backend/api/models/user.py
@@ -34,46 +34,3 @@ class User(db.Model):
 
     def __repr__(self):
         return "<User '{}'>".format(self.username)
-
-
-    @staticmethod
-    def encode_auth_token(user_id):
-        """
-        Generates the Auth Token
-        :return: string
-        """
-        try:
-            payload = {
-                'exp': datetime.datetime.utcnow() + datetime.timedelta(days=1, seconds=5),
-                'iat': datetime.datetime.utcnow(),
-                'sub': user_id
-            }
-            return jwt.encode(
-                payload,
-                key,
-                algorithm='HS256'
-            ).decode('utf-8')
-        except Exception as e:
-            return e
-
-
-    @staticmethod
-    def decode_auth_token(auth_token):
-        """
-        Decodes the auth token
-        :param auth_token:
-        :return: dict|string
-        """
-        try:
-            payload = jwt.decode(auth_token, key)
-            is_blacklisted_token = InvalidToken.check_blacklist(auth_token)
-            if is_blacklisted_token:
-                return 'Token blacklisted. Please log in again.'
-            else:
-                return {'username': payload['sub']}
-        except jwt.ExpiredSignatureError:
-            return 'Signature expired. Please log in again.'
-        except jwt.InvalidTokenError:
-            return 'Invalid token. Please log in again.'
-        except Exception:
-            return 'Unknown exception.'
